Original change: https://android-review.googlesource.com/c/platform/frameworks/support/+/3652453

+++ pdf/pdf-viewer/src/main/kotlin/androidx/pdf/view/PdfViewAccessibilityManager.kt
@@ -79,8 +79,8 @@ internal class PdfViewAccessibilityManager(
             return fastScrollPageIndicatorBackgroundDrawableId
         }
 
-        val contentX = pdfView.toContentX(x).toInt()
-        val contentY = pdfView.toContentY(y).toInt()
+        val contentX = pdfView.toContentX(x)
+        val contentY = pdfView.toContentY(y)
 
         if (!isLinksLoaded) {
             loadPageLinks()
